Program so far…
Nothing close to finished… have to change computers.
Included is a lengthy Package class which will probably have an empty
constructor and a bunch of set methods.
and a class that will read the Status file, and produce a package list
that'll eventually be configured by options the user enters in
Program.java